# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "traffic-api"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api"
    }
  ],
  "paths": {
    "/cameras": {
      "get": {
        "summary": "Get details for all cameras",
        "operationId": "listCameras",
        "tags": ["camera"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cameras",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Camera"
                  }
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/objects": {
      "get": {
        "summary": "Get details for all objects",
        "operationId": "listObjects",
        "tags": ["object"],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "description": "filter by start timestamp",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "filter by end timestamp",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Object"
                  }
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/traffic": {
      "get": {
        "summary": "Get details for all mock traffic count data",
        "operationId": "listTrafficData",
        "tags": ["traffic"],
        "responses": {
          "200": {
            "description": "List of sample traffic data points",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Traffic"
                  }
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Camera": {
        "type": "object",
        "properties": {
          "location": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "_id": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "Object": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "className": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer"
          },
          "properties": {
            "type": "object",
            "properties": {
              "c": {
                "type": "number"
              },
              "x": {
                "type": "number"
              },
              "y": {
                "type": "number"
              },
              "w": {
                "type": "number"
              },
              "h": {
                "type": "number"
              }
            }
          },
          "camera": {
            "type": "integer"
          }
        }
      },
      "Traffic": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "location": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "name": {
            "type": "string"
          },
          "volume": {
            "type": "integer"
          },
          "startTime": {
            "type": "integer"
          },
          "endTime": {
            "type": "integer"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}